+++ src/frontend/screens/Audio/AudioSavedBottomSheet.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react';
-import {BottomSheetWrapper} from '../sharedComponents/BottomSheetWrapper';
+import {BottomSheetWrapper} from '../../sharedComponents/BottomSheetWrapper';
 import {StyleSheet, View} from 'react-native';
 import {defineMessages, useIntl} from 'react-intl';
 import GreenCheck from '../images/GreenCheck.svg';
-import {HeaderText} from '../sharedComponents/Text/HeaderText';
-import {BodyText} from '../sharedComponents/Text/BodyText';
-import {Button} from '../sharedComponents/Button';
-import {NativeRootNavigationProps} from '../sharedTypes/navigation';
+import {HeaderText} from '../../sharedComponents/Text/HeaderText';
+import {BodyText} from '../../sharedComponents/Text/BodyText';
+import {Button} from '../../sharedComponents/Button';
+import {NativeRootNavigationProps} from '../../sharedTypes/navigation';
 
 const m = defineMessages({
   title: {
