+++ src/frontend/screens/Audio/shared.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import {StyleSheet} from 'react-native';
-import {WHITE, MAGENTA, DARK_GREY, BLACK} from '../lib/styles';
+import {WHITE, MAGENTA, DARK_GREY, BLACK} from '../../lib/styles';
 import {NativeStackNavigationOptions} from '@react-navigation/native-stack';
-import {CustomHeaderLeft} from './CustomHeaderLeft';
+import {CustomHeaderLeft} from '../../sharedComponents/CustomHeaderLeft';
 
 const PRIMARY_CONTROL_DIAMETER = 96;
 export const SIDE_ICON_BUTTON_WIDTH = 36;
